+++ HPSDetection.h
@@ -1,3 +1,3 @@
-float* HarmonicProductSpectrum(double** AudioData, int size, int dftBlocksize,
+float* HarmonicProductSpectrum(float** AudioData, int size, int dftBlocksize,
 			       int hpsOvr, int fftSize, int samplerate);
 float BinToFreq(int bin, int fftSize, int samplerate);
